CI: Switch PR workflow to praktika
